requirements.txt
chromadb==0.5.5
datasets==2.21.0
langchain==0.2.15
langchain-openai==0.1.23
langchain-community==0.2.15
open_clip_torch==2.26.1

